refactor(absences): split actions into focused modules
